dynamic_led_display_prod/requirements2.txt
aioserial==1.3.1
alabaster==0.7.13
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
atomicwrites==1.4.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.10.1
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.4
curio==1.6
cycler==0.12.1
daphne==4.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.6
django-admin-interface==0.26.1
django-admin-rangefilter==0.11.2
django-allauth==0.57.0
django-ash==0.0.7
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-crontab==0.7.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.14.0
docrepr==0.2.0
docutils==0.18.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
fonttools==4.43.1
fqdn==1.5.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.26.0
ipyparallel==8.6.1
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyserial==3.5
pyserial-asyncio==0.6
pytest==6.2.5
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-memcached==1.59
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
scipy==1.11.3
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
terminado==0.17.1
testpath==0.6.0
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
trio==0.22.2
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
watchdog==2.1.3
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websocket-server==0.6.4
websockets==11.0.3
widgetsnbextension==4.0.9
windrose==1.9.0
XlsxWriter==3.1.9
zope.interface==6.1
